+++ NEWS.md
@@ -1,8 +1,13 @@
 # sfdep (development version)
 
+- `spatial_gini()` is a new global measure
+- `euclidean_median()` is added for identifying the euclidean median of points for point pattern analysis. 
+  - Implemented using `pracma::geo_median()`
+    - `pracma` is now added as a suggested package
+- `center_mean()` and ``center_median()` are added for point pattern analysis
+- `tally_jc()` returns join count matrix via `spdep::joincount.multi()` as a data.frame object
 - `st_complete_nb()` creates a complete graph of neighbors 
 - set operations for neighbor lists `nb_union()`, `nb_intersect()`, `nb_setdiff()`
-- `nb_as_matrix()` and `wt_as_matrix()` to convert neighbor or weights lists to matrix representation
 - `st_block_nb()` is introduced to create neighbor contiguity based on spatial regimes
   - inspired by https://pysal.org/libpysal/generated/libpysal.weights.block_weights.html
 - `critical_threshold()` gains argument `k` to identify minimum number of neighbors when calculating threshold